requirements.txt
streamlit>=1.40.0
ollama>=0.4.0
gitpython>=3.1.0
